// frontend/src/api/generated/knowledge-client/api.ts
export const DatasetAclEnum = {
  DatasetActionUnspecified: "DATASET_ACTION_UNSPECIFIED",
  DatasetRead: "DATASET_READ",
  DatasetWrite: "DATASET_WRITE",
  DatasetUpload: "DATASET_UPLOAD",
} as const;

export type DatasetAclEnum =
  (typeof DatasetAclEnum)[keyof typeof DatasetAclEnum];